Refactored some structure
