+++ src/Command/CompareCommand.php
@@ -2,6 +2,7 @@
 //----------------------------------------------------------------------------------------------------------------------
 namespace SetBased\DataSync\Command;
 
+use SetBased\DataSync\Command\Compare\CompareMasterData;
 use SetBased\Stratum\Style\StratumStyle;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
