+++ src/Command/Dump/Components/Node.php
@@ -1,6 +1,6 @@
 <?php
 //----------------------------------------------------------------------------------------------------------------------
-namespace SetBased\DataSync;
+namespace SetBased\DataSync\Command\Dump\Components;
 
 //----------------------------------------------------------------------------------------------------------------------
 /**
